Moved authentication classes into a new sub-package
