+++ app/src/main/AndroidManifest.xml
@@ -62,8 +62,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LoginActivity"></activity>
-        <activity android:name=".OAuthActivity"></activity>
+        <activity android:name=".authentication.LoginActivity"></activity>
+        <activity android:name=".authentication.OAuthActivity"></activity>
     </application>
 
 </manifest>
