+++ app/src/main/java/com/dankideacentral/dic/authentication/LoginActivity.java
@@ -1,17 +1,20 @@
-package com.dankideacentral.dic;
+package com.dankideacentral.dic.authentication;
 
-import android.content.Context;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 // Copied from the tutorial
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.net.Uri;
 import android.os.AsyncTask;
 import android.preference.PreferenceManager;
 import android.view.View;
 import android.widget.Button;
+
+import com.dankideacentral.dic.MainActivity;
+import com.dankideacentral.dic.R;
+import com.dankideacentral.dic.TwitterUtil;
+
 import twitter4j.auth.RequestToken;
 
 public class LoginActivity extends AppCompatActivity {
