+++ app/src/main/java/com/dankideacentral/dic/authentication/OAuthWebViewFragment.java
@@ -1,4 +1,4 @@
-package com.dankideacentral.dic;
+package com.dankideacentral.dic.authentication;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -10,6 +10,9 @@ import android.view.ViewGroup;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
+
+import com.dankideacentral.dic.MainActivity;
+import com.dankideacentral.dic.R;
 
 public class OAuthWebViewFragment extends Fragment {
     private WebView webView;
